move stuff into shared places, so we can reuse it for ADTs
